healthyrizz_vps_deploy/requirements.txt
# Flask and Extensions
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Admin==1.6.1
Flask-Babel==4.0.0
Flask-Caching==2.1.0
Flask-Compress==1.14
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-Session==0.6.0
Flask-CSP==0.10

# Database
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# PDF Generation
reportlab==4.1.0
WeasyPrint==60.2
PyPDF2==3.0.1

# Email and Notifications
pywebpush==1.14.0
sendgrid==6.11.0
twilio==8.12.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==42.0.2
python-dotenv==1.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
Pillow==10.2.0
gunicorn==21.2.0
eventlet==0.35.2
gevent==24.2.1

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Production
# uwsgi==2.0.24
supervisor==4.2.5

# Monitoring and Logging
sentry-sdk==1.40.4
python-json-logger==2.0.7

# API and Integration
stripe==7.11.0
paypalrestsdk==1.13.1
boto3==1.34.34
google-cloud-storage==2.14.0

# Caching and Performance
cachetools==5.3.3
ujson==5.9.0
orjson==3.9.15

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Optional Dependencies
# Uncomment if needed
# elasticsearch==8.11.1
# celery==5.3.6
# flower==2.0.1
# prometheus-client==0.19.0
# grafana-api-client==3.0.0 